5. Offline RAG/requirements.txt
streamlit
embedchain
streamlit-chat
pillow
pytesseract
